Added interface for ConsortiumService
